# apps/web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniFr Planner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/account/plans": {
      "get": {
        "tags": ["account"],
        "summary": "List Plans",
        "operationId": "list_plans_api_v1_account_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          }
        },
        "security": [
          {
            "PrivateSession": []
          }
        ]
      }
    },
...
    "/api/v1/account": {
      "delete": {
        "tags": ["account"],
        "summary": "Delete Account",
        "operationId": "delete_account_api_v1_account_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivateError"
                }
              }
            }
          }
        },
        "security": [
          {
            "PrivateSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Password": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["password"],
        "title": "Password"
      },
      "PlanList": {
        "properties": {
          "plans": {
            "items": {
              "$ref": "#/components/schemas/SavedPlan"
            },
            "type": "array",
            "title": "Plans"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["plans"],
        "title": "PlanList"
      },
...
      "PrivateError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["detail"],
        "title": "PrivateError"
      },
...
      "SavedPlan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          },
          "conflictOf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflictof"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "revision", "snapshot"],
        "title": "SavedPlan"
      },
...
    },
    "securitySchemes": {
      "PrivateSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-unifr_session"
      }
    }
  }
}